refactor: Refactor entire frontend part     1. 컴포넌트를 프레젠테이션부분과 컨테이너부분으로 분리     2. 폴더구조를 프레젠테이션와 컨테이너에 맞춰 정리     3. 컬럼부분 CSS 수정 (title 부분)
