+++ frontend/src/App.js
@@ -1,8 +1,7 @@
 import React, { useEffect } from 'react';
 import styled from "styled-components";
 import HeaderContainer from './containers/header/HeaderContainer';
-
-import Column from './components/Seong/Column';
+import MainContainer from './containers/main/MainContainer';
 
 const StyledAppWrapper = styled.div`
     width: 1440px;
@@ -11,18 +10,6 @@ const StyledAppWrapper = styled.div`
     position: relative;
 `;
 
-const list = [
-    {
-        title: '대충제목',
-        body: '대충내용',
-        author: 'web',
-    },
-    {
-        title: '제목',
-        body: '대충ㅇㅇㅇㅇㅇㅇㅇㅇㅇ내용',
-        author: 'web',
-    },
-];
 
 const App = () => {
     useEffect(() => (document.body.style = 'background: #F5F5F7;'), []);
@@ -30,7 +17,8 @@ const App = () => {
     return (
         <StyledAppWrapper>
             <HeaderContainer />
-            <Column title="Column title" list={list} />
+            <MainContainer />
+            {/* <Column title="Column title" list={list} /> */}
         </StyledAppWrapper>
     );
 };
