+++ frontend/src/components/main/Main.js
@@ -1,0 +1,12 @@
+import React from "react";
+import styled from "styled-components";
+
+const StyledMainUl = styled.ul`
+    display: flex;
+`;
+
+const Main = ({children}) => {
+    return <StyledMainUl>{children}</StyledMainUl>
+};
+
+export default Main;
